:bug: :: .충돌 해결
